+++ packages/react/src/index.d.ts
@@ -14,6 +14,7 @@ export { Notification } from "./components/notification/Notification";
 export { Snackbar, SnackbarStack } from "./components/snackbar/Snackbar";
 export { StatusBar } from "./components/status-bar/StatusBar";
 export { Stat } from "./components/stat/Stat";
+export { Avatar } from "./components/avatar/Avatar";
 export {
   AreaChart,
   BarChart,